requirements.txt
fastapi
uvicorn
python-dotenv
transformers
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
git+https://github.com/openai/whisper.git
numpy==1.26.4
python-multipart